Polissage de certains features

+++ admin/views/listeUtilisateursView.php
@@ -63,10 +63,12 @@ if ($_SESSION["estConnecte"] != true) {
                                                     </a>
                                                 </td>
                                                 <td>
-                                                    
-                                                <a href="supprimer-utilisateur.php?id=<?php echo $utilisateur['id']; ?>">
-                                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"><path d="M6 2l2-2h4l2 2h4v2H2V2h4zM3 6h14l-1 14H4L3 6zm5 2v10h1V8H8zm3 0v10h1V8h-1z"/></svg>
+                                                
+                                                <?php if($utilisateur['id'] != 3) { ?>
+                                                    <a href="supprimer-utilisateur.php?id=<?php echo $utilisateur['id']; ?>">
+                                                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"><path d="M6 2l2-2h4l2 2h4v2H2V2h4zM3 6h14l-1 14H4L3 6zm5 2v10h1V8H8zm3 0v10h1V8h-1z"/></svg>
                                                     </a>
+                                                <?php } ?>
                                                 </td>
                                             </tr>
                                         <?php } ?>
